requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.3.4
python-multipart==0.0.9
requests==2.32.3
Pillow==10.4.0
qdrant-client==1.11.1
numpy==1.26.4
